+++ app/libraries/Theme/Setups/MetaBoxes/Featured.php
@@ -48,7 +48,7 @@ final class Featured extends AbstractMetaBox
             $fields[] = [
                 'id' => $utilities->awesomePosts->id(),
                 'type' => 'checkbox',
-                'label' => \esc_html__('Add to Awesome Posts', 'my-theme'),
+                'label' => \esc_html__('Add to Awesome Posts', 'jentil-theme'),
             ];
         }
 
@@ -56,7 +56,7 @@ final class Featured extends AbstractMetaBox
             return [
                 'id' => $this->id,
                 'context' => $this->context,
-                'title' => \esc_html__('Featured', 'my-theme'),
+                'title' => \esc_html__('Featured', 'jentil-theme'),
                 'priority' => 'default',
                 'callbackArgs' => ['__block_editor_compatible_meta_box' => true],
                 'fields' => $fields,